requirements.txt
Django==1.8
Markdown==2.6.6
argparse==1.2.1
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.9
django-filter==0.13.0
djangorestframework==3.3.3
ipdb==0.10.0
ipython==4.2.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.2
ptyprocess==0.5.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
wsgiref==0.1.2
